requirements.txt
aiohttp>=2.1,<2.3
discord.py>=0.16.12
